+++ src/main/java/software/xdev/pmd/config/PluginConfigurationBuilder.java
@@ -1,5 +1,6 @@
 package software.xdev.pmd.config;
 
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Objects;
 import java.util.SortedSet;
@@ -23,19 +24,23 @@ public final class PluginConfigurationBuilder
 	private boolean showSuppressedWarnings;
 	private boolean useCacheFile;
 	private ScanScope scanScope;
+	private SortedSet<PatternContainer> projectRelativeFileExclusions;
 	private SortedSet<ConfigurationLocation> locations;
 	private SortedSet<String> activeLocationIds;
+	private boolean importSettingsFromMaven;
 	
 	public PluginConfigurationBuilder(final Project project)
 	{
 		this.showSuppressedWarnings = true;
 		this.useCacheFile = true;
 		this.scanScope = ScanScope.getDefaultValue();
+		this.projectRelativeFileExclusions = Collections.emptySortedSet();
 		this.locations = BundledConfig.getAllBundledConfigs()
 			.stream()
 			.map(bc -> configurationLocationFactory(project).create(bc, project))
 			.collect(Collectors.toCollection(TreeSet::new));
 		this.activeLocationIds = Collections.emptySortedSet();
+		this.importSettingsFromMaven = false;
 	}
 	
 	public PluginConfigurationBuilder(final PluginConfiguration copyFrom)
@@ -44,8 +49,10 @@ public final class PluginConfigurationBuilder
 		this.showSuppressedWarnings = copyFrom.showSuppressedWarnings();
 		this.useCacheFile = copyFrom.useCacheFile();
 		this.scanScope = copyFrom.scanScope();
+		this.projectRelativeFileExclusions = copyFrom.projectRelativeFileExclusions();
 		this.locations = copyFrom.locations();
 		this.activeLocationIds = copyFrom.activeLocationIds();
+		this.importSettingsFromMaven = copyFrom.importSettingsFromMaven();
 	}
 	
 	public static PluginConfiguration copy(@NotNull final PluginConfiguration source)
@@ -80,6 +87,26 @@ public final class PluginConfigurationBuilder
 		return this;
 	}
 	
+	public PluginConfigurationBuilder withProjectRelativeFileExclusionsRaw(
+		final Collection<String> projectRelativeFileExclusions)
+	{
+		if(projectRelativeFileExclusions == null)
+		{
+			return this.withProjectRelativeFileExclusions(null);
+		}
+		return this.withProjectRelativeFileExclusions(projectRelativeFileExclusions.stream()
+			.map(PatternContainer::tryCreate)
+			.filter(Objects::nonNull)
+			.collect(Collectors.toCollection(TreeSet::new)));
+	}
+	
+	public PluginConfigurationBuilder withProjectRelativeFileExclusions(
+		final SortedSet<PatternContainer> projectRelativeFileExclusions)
+	{
+		this.projectRelativeFileExclusions = projectRelativeFileExclusions;
+		return this;
+	}
+	
 	public PluginConfigurationBuilder withActiveLocationIds(@NotNull final SortedSet<String> newActiveLocationIds)
 	{
 		this.activeLocationIds = newActiveLocationIds;
@@ -98,6 +125,12 @@ public final class PluginConfigurationBuilder
 		return this;
 	}
 	
+	public PluginConfigurationBuilder withImportSettingFromMaven(final boolean importSettingsFromMaven)
+	{
+		this.importSettingsFromMaven = importSettingsFromMaven;
+		return this;
+	}
+	
 	public PluginConfiguration build()
 	{
 		return new PluginConfiguration(
@@ -105,12 +138,16 @@ public final class PluginConfigurationBuilder
 			this.showSuppressedWarnings,
 			this.useCacheFile,
 			this.scanScope,
+			Collections.unmodifiableSortedSet(Objects.requireNonNullElseGet(
+				this.projectRelativeFileExclusions,
+				TreeSet::new)),
 			Collections.unmodifiableSortedSet(Objects.requireNonNullElseGet(this.locations, TreeSet::new)),
 			this.activeLocationIds != null
 				? this.activeLocationIds.stream()
 				.filter(Objects::nonNull)
 				.collect(Collectors.toCollection(TreeSet::new))
 				: new TreeSet<>(),
+			this.importSettingsFromMaven,
 			new PluginConfiguration.Cache());
 	}