+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 1.2.0
+* Make it possible to add (project-wide) exclusions
+* Add support for importing settings from Maven PMD plugin (best-effort)
 * Add shortcuts for jumping to previous/next result/violation
 * Add a "scan entire project" button #59
 * Only show "Enable automatic build" hint on startup when projects have PMD enabled